add Procedures that read .wav files and add timer and updating playerwon Procedures
the Procedures StartSound, read, br, PrintAX made by:
https://gist.github.com/w00tious/0334d66f73b0fc9e604e6b67878fd57b the
Procedures read wav file and play it. the Procedures ResetTimer,
TestTimersMinutes, TestTimersSeconds, RearrangeTimerNombers, StartTimer,
EndTimer, PlayerTimeOutPut, SoundEnd are in charge on the timer. the
timer have 2 uses 1 to manage the sound and the other use is the count
how mach time its take to finish a level